+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Inconsistent spacing in configuration UI
 - Invalid revision when cancelling edit mode without changes
 - Set prompt library defaults on first language model added
+- Unable to cancel changes with active editor closed
 
 ## [2024.10.1] - 2024-09-05